+++ rocketgram/executors/aiohttp.py
@@ -18,19 +18,17 @@ from ..types import InputFile
 from ..update import Update
 from ..version import version
 
-if TYPE_CHECKING:
-    from ..bot import Bot
-
-json_encoder = json.dumps
-json_decoder = json.loads
-
 try:
     import ujson
 
     json_encoder = ujson.dumps
     json_decoder = ujson.loads
-except ModuleNotFoundError:
-    pass
+except ImportError:
+    json_encoder = json.dumps
+    json_decoder = json.loads
+
+if TYPE_CHECKING:
+    from ..bot import Bot
 
 logger = logging.getLogger('rocketgram.executors.aiohttp')